+++ android/history.md
@@ -3,6 +3,7 @@
 ## 2020-05-07 13.0.6210 stable
 * Bug fix:
   * Back button dismisses system keyboard (#3094)
+  * Fix system keyboard alignment (#3097)
 
 ## 2020-04-20 13.0.6209 stable
 * Bug fix: